requirements.txt
-i https://pypi.org/simple

django==2.2.20

djangorestframework==3.12.4

psycopg2==2.8.6

pytz==2021.1

sqlparse==0.4.1; python_version >= '3.5'

